Arreglo archivo pom.xml, Creación de DTOs, Mapper y excepciones (BackEnd Actualizado)

+++ src/main/java/com/example/Restaurante/Modelo/Platos.java
@@ -5,17 +5,16 @@ import lombok.Data;
 
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Entity
 @Data
 @Table(name = "platos")
 public class Platos implements Serializable {
-
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "platos_generator")
     @SequenceGenerator(name = "platos_generator", sequenceName = "platos_seq", allocationSize = 1)
     private Long id;
     private String nombre;
-    private int valor;
-
+    private BigDecimal valor;
 }
